Relocated client/utils.js into client/utils/index.js, so it was possible to use 'import {} from "utils"'.
The change simplifies codebase and improves maintainability.
